Refactor EntrypointParams into Entrypoint
