+++ testing/test_actors/actors/fil-upgrade-actor/Cargo.toml
@@ -5,8 +5,8 @@ edition = "2021"
 publish = false
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
-fvm_sdk = { version = "4.0.0-alpha.3", path = "../../../../sdk" }
-fvm_shared = { version = "4.0.0-alpha.3", path = "../../../../shared" }
+fvm_sdk = { version = "4.0.0-alpha.4", path = "../../../../sdk" }
+fvm_shared = { version = "4.0.0-alpha.4", path = "../../../../shared" }
 fvm_ipld_encoding = { version = "0.4.0", path = "../../../../ipld/encoding" }
 cid = { workspace = true }
 serde = { version = "1.0.164", features = ["derive"] }
